web3_nocode_backend/requirements.txt
fastapi
uvicorn
motor
beanie
redis
python-dotenv
web3
pydantic[email]
